Mind
The Way bears sensation,
Sensation bears memory,
Sensation and memory bear abstraction,
And abstraction bears all the world;
Each thing in the world bears feeling and doing,
And, imbued with mind, harmony with the Way.
As others have taught, so do I teach,
"Who loses harmony opposes nature";
This is the root of my teaching.
		-- Lao Tse, "Tao Te Ching"