Better doing NNTP from scratch.
Developer of go-nntp is just a former java deleloper. It has keep all the bad habits of Java, plus the bad habits of C. Code is unreadable , hard to support , hard to debug, hard to reuse, and no documentation. Is a good example of how to DON'T write code, regardless it works or not. Better if I write the protocol by myself.